pyproject.toml
[tool.poetry]
name = "test-job-o-complex"
version = "0.1.0"
description = ""
authors = ["Lewbertsklol <lewbertsklol@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
